requirements.txt
selenium==3.141.0
msedge-selenium-tools==3.141.3
requests==2.25.1
allure-pytest==2.9.43
allure-python-commons==2.9.43
pytest==6.2.4
PyYAML==5.4.1
Faker==8.2.1
tenacity==7.0.0
multidict==5.1.0
pytest-timeout==1.4.2
pylint==2.8.2
webdriver-manager==3.4.2
selenium-wire==4.5.3
pytest-xdist==2.3.0
jsonschema==4.1.2
dataclasses_json==0.5.6
rstr==3.0.0
PyPDF2==1.26.0
pyautogui==0.9.53